+++ src/components/Modal/SimpleModal.vue
