+++ src/components/Modal/minimalEditor/MiniEditor.vue
