+++ src/components/extensions/recipe/recipeRule.ts
@@ -3,7 +3,7 @@ import {
   mergeAttributes
 } from '@tiptap/core';
 import { VueNodeViewRenderer } from '@tiptap/vue-3'
-import RecipeNodeVue from '../../node/recipe/RecipeNode.vue'
+import RecipeNodeVue from '../../nodeVue/recipe/RecipeNode.vue'
 
 export interface RecipeRuleOptions {
   // HTMLAttributes: Record<string, any>;
