+++ src/components/nodeVue/condition/ConditionNode.vue
