+++ src/components/nodeVue/recipe/RecipeNode.vue
