--- src/testFolder/index.vue
@@ -1,53 +1,0 @@
-<template>
-  <editor-content :editor="editor" />
-</template>
-
-<script>
-import StarterKit from '@tiptap/starter-kit'
-import { Editor, EditorContent } from '@tiptap/vue-3'
-
-import VueComponent from './Extension.js'
-
-export default {
-  components: {
-    EditorContent,
-  },
-
-  data() {
-    return {
-      editor: null,
-    }
-  },
-
-  mounted() {
-    this.editor = new Editor({
-      extensions: [
-        StarterKit,
-        VueComponent,
-      ],
-      content: `
-        <p>
-          This is still the text editor you’re used to, but enriched with node views.
-        </p>
-        <vue-componen count="99"></vue-componen>
-        <p>
-          Did you see that? That’s a Vue component. We are really living in the future.
-        </p>
-      `,
-    })
-  },
-
-  beforeUnmount() {
-    this.editor.destroy()
-  },
-}
-</script>
-
-<style lang="scss">
-/* Basic editor styles */
-.tiptap {
-  > * + * {
-    margin-top: 0.75em;
-  }
-}
-</style>
